requirements.txt
rumps
pynput
sounddevice
numpy
openai-whisper
pyobjc-core
pyobjc-framework-Cocoa
pyperclip
